Do path prediction for non-dead nodes even when the local player has died.
Previously, we would stop doing path prediction for all nodes if the
local player died. This didn't make sense because the other still alive
players would obviously still be moving, and the local player should
see this movement whether they are dead or not.

Additionally, this changes the previous behaviour of "send game state
updates as quickly as possible even when the player is not playing,
forever" to remove the "as quickly as possible" part. The previous
behaviour worked, but would noticeably slow down the browser by
sending extremely frequent messages.